refactor: move handleDownload to utils, add selector for pokemonSelectionSlice
