+++ src/components/SelectionFlyout/SelectionFlyout.tsx
@@ -2,6 +2,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { clearAllSelections } from '../store/pokemonSelectionSlice';
 import type { RootState } from '../../utils/store';
 import './SelectionFlyout.css';
+import { handleDownload } from '../../utils/common';
 
 export const SelectionFlyout = () => {
   const dispatch = useDispatch();
@@ -9,23 +10,6 @@ export const SelectionFlyout = () => {
     (state: RootState) => state.pokemonSelection.selectedPokemons
   );
   const selectedCount = Object.keys(selectedPokemons).length;
-
-  const handleDownload = () => {
-    const csvContent = [
-      'ID,Name,Types,Abilities,URL',
-      ...Object.values(selectedPokemons).map(
-        (p) =>
-          `${p.id},${p.name},${p.types.join('|')},${p.abilities.join('|')},${p.url}`
-      ),
-    ].join('\n');
-
-    const blob = new Blob([csvContent], { type: 'text/csv' });
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement('a');
-    link.href = url;
-    link.download = `${selectedCount}_pokemons.csv`;
-    link.click();
-  };
 
   if (selectedCount === 0) return null;
 
@@ -38,7 +22,10 @@ export const SelectionFlyout = () => {
       >
         Unselect all
       </button>
-      <button className="flyout-button download" onClick={handleDownload}>
+      <button
+        className="flyout-button download"
+        onClick={() => handleDownload(selectedPokemons, selectedCount)}
+      >
         Download CSV
       </button>
     </div>
