+++ src/utils/common.ts
@@ -15,3 +15,23 @@ export const prepareData = (pokemonData: PokemonDetailResponse): Pokemon => ({
   weight: pokemonData.weight,
   stats: pokemonData.stats,
 });
+
+export const handleDownload = (
+  selectedPokemons: Pokemon[],
+  selectedCount: number
+) => {
+  const csvContent = [
+    'ID,Name,Types,Abilities,URL',
+    selectedPokemons.map(
+      (p) =>
+        `${p.id},${p.name},${p.types.join('|')},${p.abilities.join('|')},${p.url}`
+    ),
+  ].join('\n');
+
+  const blob = new Blob([csvContent], { type: 'text/csv' });
+  const url = URL.createObjectURL(blob);
+  const link = document.createElement('a');
+  link.href = url;
+  link.download = `${selectedCount}_pokemons.csv`;
+  link.click();
+};
